Rename AoE module so it's obvious
