+++ src/parser/jobs/mnk/modules/index.ts
@@ -1,9 +1,9 @@
 import {ActionTimeline} from './ActionTimeline'
+import {AoE} from './AoE'
 import {Demolish} from './Demolish'
 import {Fists} from './Fists'
 import {Forms} from './Forms'
 import {GreasedLightning} from './GreasedLightning'
-import {AoE} from './MultiHitSkills'
 import {OGCDDowntime} from './OGCDDowntime'
 import {PerfectBalance} from './PerfectBalance'
 import {RiddleOfFire} from './RiddleOfFire'
@@ -13,11 +13,11 @@ import {Weaving} from './Weaving'
 
 export default [
 	ActionTimeline,
+	AoE,
 	Demolish,
 	Fists,
 	Forms,
 	GreasedLightning,
-	AoE,
 	OGCDDowntime,
 	PerfectBalance,
 	RiddleOfFire,
